move virtual tank registry into own package
