+++ src/main/java/gregtech/api/util/virtualregistry/VirtualTankRegistry.java
@@ -1,6 +1,8 @@
-package gregtech.api.util;
+package gregtech.api.util.virtualregistry;
 
 import gregtech.api.GTValues;
+
+import gregtech.api.util.GTLog;
 
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.World;
